requirements.txt
Django>=3.0,<4.0
django-rest-framework==0.1.0
psycopg2-binary==2.8.6
requests==2.24.0
gunicorn==20.0.4
discord.py==1.4.1
trueskill==0.4.5
sentry-sdk==0.19.3
aiohttp-json-rpc==0.13.2
